+++ apps/koku-ui-ros/src/api/queries/rosQuery.ts
@@ -13,13 +13,16 @@ interface RosGroupBys {
 }
 
 export interface RosQuery extends utils.Query {
+  after?: string;
   category?: string;
   delta?: string;
+  engine?: string;
   filter?: RosFilters;
   group_by?: RosGroupBys;
   limit?: number;
   offset?: number;
   order_by?: any;
+  term?: string;
 }
 
 // filter_by props are converted and returned with logical OR/AND prefix